+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 1.0.1-SNAPSHOT
+# 1.0.2-SNAPSHOT
+
+* Improvement: Apply the Java plugin instead of the Base plugin. Run `compileJava` between `compileLanguage` and `archiveLanguage`.
+* Improvement: Configure `compileJava` to use the ECJ compiler (instead of default JDK compiler) to compile the huge mess of Spoofax-generated Java.
+
+# 1.0.1
 
 * Bugfix: Add repositories that are needed to build the plugin (Metaborg, Pluto Build, Sugar Lang, usethesource) and replace mavenCentral by jcenter.
 * Improvement: Add a test dependency on JUnit 5 (Jupiter API & engine), update and move the example projects to the test resources directory, and configure Gradle to run the tests.